pyproject.toml
[project]
name = "rookiecard-hold-agent"
version = "0.1.0"
description = "Hold Agent — daily portfolio health monitor for NBA rookie cards, with eBay price scraping and MCP integration for messaging platforms (WeChat, Telegram, Discord via Hermes)."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "eh3323" }]
dependencies = [
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "curl_cffi>=0.7.0",
    "playwright>=1.52.0",
    "playwright-stealth>=1.0.6",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
rookiecard-mcp = "rookiecard.mcp.server:main"
rookiecard-ebay = "rookiecard.scrapers.ebay:__main__"

[tool.setuptools.packages.find]
include = ["rookiecard*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
